Create initial Home and MyMentors pages with navigation
