+++ App.js
@@ -1,20 +1,51 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
+import { SafeAreaProvider } from 'react-native-safe-area-context';
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { Button, lightColors, createTheme, ThemeProvider, Avatar } from '@rneui/themed';
+import Home from './screens/Home';
+import MyMentors from './screens/MyMentors';
+
+// const theme = createTheme({
+//   // Colors
+//   lightColors: {
+//     primary: '#e7e7e8',
+//   },
+//   darkColors: {
+//     primary: '#000',
+//   },
+//   mode: 'light',
+//   // Component styles
+//   components: {
+//     Button: {
+//       raised: true,
+//       color: "primary",
+//     },
+//   },
+// });
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Text>Open up App.js to start working on your app!</Text>
-      <StatusBar style="auto" />
-    </View>
+    <ThemeProvider>
+      <SafeAreaProvider>
+        <NavigationContainer>
+          <Stack.Navigator
+            // screenOptions={{
+            //   headerStyle: { backgroundColor: "green" },
+            //   headerTintColor: "white",
+            //   headerTitleStyle: { fontSize: 30 },
+            // }}
+          >
+          <Stack.Screen name="Home" component={Home} />
+          <Stack.Screen name="MyMentors" component={MyMentors} />
+          </Stack.Navigator>
+        </NavigationContainer>
+      </SafeAreaProvider>
+    </ThemeProvider>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+
