+++ screens/MyMentors.js
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+export default function MyMentors() {
+  return (
+    <View>
+      <Text>MyMentors</Text>
+    </View>
+  )
+}
